refactor: rename hercule to hercule-core, add hercule package for compatability
