+++ packages/hercule-core/.npmignore
@@ -1,0 +1,2 @@
+src
+test
