+++ packages/hercule-core/package.json
@@ -1,15 +1,12 @@
 {
-  "name": "hercule",
+  "name": "hercule-core",
   "version": "4.1.1",
   "description": "Markdown, API Blueprint and string transclusion",
   "main": "./lib/hercule",
-  "bin": {
-    "hercule": "./bin/hercule"
-  },
   "engines": {
     "node": ">=4.0.0"
   },
-  "repository": "https://github.com/jamesramsay/hercule/tree/master/packages/hercule",
+  "repository": "https://github.com/jamesramsay/hercule/tree/master/packages/hercule-core",
   "keywords": [
     "markdown",
     "transclude",
@@ -30,7 +27,16 @@
   },
   "homepage": "https://github.com/jamesramsay/hercule",
   "dependencies": {
-    "hercule-cli": "^4.1.1",
-    "hercule-core": "^4.1.1"
+    "async": "^2.1.4",
+    "clone-regexp": "^1.0.0",
+    "duplexer3": "^0.1.4",
+    "get-stream": "^3.0.0",
+    "got": "^8.0.0",
+    "isstream": "^0.1.2",
+    "left-split": "^1.0.0",
+    "lodash": "^4.0.0",
+    "source-map": "^0.6.0",
+    "through2": "^2.0.0",
+    "through2-get": "^0.1.0"
   }
 }
