+++ packages/hercule-core/src/config.js
