+++ packages/hercule-core/src/grammar/link.pegjs
