+++ packages/hercule-core/src/resolver.js
