+++ packages/hercule-core/src/trim.js
