+++ packages/hercule-core/test/.eslintrc.json
