+++ packages/hercule-core/test/integration/_mock.js
