+++ packages/hercule-core/test/units/transcludeString.js
