+++ packages/hercule-core/test/units/trim.js
