+++ packages/hercule/bin/hercule.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+'use strict';
+
+require('hercule-cli/bin/hercule');
